+++ src/hooks/useBoolean.ts
@@ -1,6 +1,6 @@
 import { useCallback, useState } from 'react';
 
-export const useToggle = (initialValue = false) => {
+export const useBoolean = (initialValue: boolean) => {
   const [value, setValue] = useState(initialValue);
 
   const toggle = useCallback(() => {